+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/asplanned/mapper/AssetAsPlannedResponseMapper.java
@@ -46,7 +46,7 @@ public class AssetAsPlannedResponseMapper extends AssetBaseResponseMapper {
                         asset.getParentRelations().stream()
                                 .map(AssetAsPlannedResponseMapper::from)
                                 .toList())
-                .underInvestigation(asset.isUnderInvestigation())
+                .underInvestigation(asset.isInInvestigation())
                 .activeAlert(asset.isActiveAlert())
                 .qualityType(
                         from(asset.getQualityType())