+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/asbuilt/repository/AssetAsBuiltRepositoryImpl.java
@@ -40,6 +40,7 @@ import java.util.List;
 
 import static org.apache.commons.collections4.ListUtils.emptyIfNull;
 import static org.eclipse.tractusx.traceability.common.repository.EntityNameMapper.toDatabaseName;
+import static org.eclipse.tractusx.traceability.common.repository.SqlUtil.constructLikeWildcardQuery;
 
 @RequiredArgsConstructor
 @Component
@@ -85,9 +86,9 @@ public class AssetAsBuiltRepositoryImpl implements AssetAsBuiltRepository {
     }
 
     @Override
-    public List<String> getFieldValues(String fieldName, Long resultLimit) {
+    public List<String> getFieldValues(String fieldName, String startWith, Long resultLimit) {
         String databaseFieldName = toDatabaseName(fieldName);
-        String getFieldValuesQuery = "SELECT DISTINCT " + databaseFieldName + " FROM assets_as_built ORDER BY " + databaseFieldName + " ASC LIMIT :resultLimit";
+        String getFieldValuesQuery = "SELECT DISTINCT " + databaseFieldName + " FROM assets_as_built" + constructLikeWildcardQuery(databaseFieldName, startWith) + " ORDER BY " + databaseFieldName + " ASC LIMIT :resultLimit";
         return entityManager.createNativeQuery(getFieldValuesQuery, String.class)
                 .setParameter("resultLimit", resultLimit)
                 .getResultList();